% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{A linear regression package}
\description{
A linear regression package
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula.}

\item{\code{data}}{data.frame.}

\item{\code{dta}}{character.}

\item{\code{frmla}}{formula.}

\item{\code{X}}{matrix.}

\item{\code{y}}{matrix.}

\item{\code{Regressions_coefficients}}{matrix.}

\item{\code{The_fitted_values}}{matrix.}

\item{\code{The_residuals}}{matrix.}

\item{\code{The_degrees_of_freedom}}{numeric.}

\item{\code{The_residual_variance}}{numeric.}

\item{\code{The_variance_of_the_regression_coefficients}}{matrix.}

\item{\code{The_t_values}}{matrix.}
}}


\examples{

linregtest <- linreg$new(Petal.Length~Species, data = iris)
linregtest$print()
linregtest$plot()
linregtest$summary()
}
